Move dashboard to a V1 dir
