+++ src/Router.tsx
@@ -1,9 +1,12 @@
 import { HashRouter, Route, Switch } from 'react-router-dom'
 import { Redirect, useParams } from 'react-router'
-import Dashboard from 'components/Dashboard'
+import DashboardV1 from 'components/DashboardV1'
+import DashboardV2 from 'components/DashboardV1' // TODO update to V2
 import Landing from 'components/Landing'
 import Create from 'components/Create'
 import Projects from 'components/Projects'
+import UserProviderV1 from 'providers/UserProviderV1'
+import UserProviderV2 from 'providers/UserProviderV1'
 
 function CatchallRedirect() {
   const route = useParams<{ route: string }>()['route']
@@ -26,8 +29,15 @@ export default function Router() {
         <Route path="/projects">
           <Projects />
         </Route>
+        <Route path="/p/:handle(.+\.eth)">
+          <UserProviderV2>
+            <DashboardV2 />
+          </UserProviderV2>
+        </Route>
         <Route path="/p/:handle">
-          <Dashboard />
+          <UserProviderV1>
+            <DashboardV1 />
+          </UserProviderV1>
         </Route>
         <Route path="/:route">
           <CatchallRedirect />
