+++ src/components/Create/index.tsx
@@ -5,7 +5,7 @@ import { Trans } from '@lingui/macro'
 import { Button, Col, Drawer, DrawerProps, Row, Space } from 'antd'
 import { useForm } from 'antd/lib/form/Form'
 import Modal from 'antd/lib/modal/Modal'
-import Project from 'components/Dashboard/Project'
+import Project from 'components/DashboardV1/Project'
 import { NetworkContext } from 'contexts/networkContext'
 import { ProjectContext, ProjectContextType } from 'contexts/projectContext'
 import { ThemeContext } from 'contexts/themeContext'
