+++ src/components/FundingCycle/QueuedFundingCycle.tsx
@@ -3,7 +3,7 @@ import { ProjectContext } from 'contexts/projectContext'
 import { ThemeContext } from 'contexts/themeContext'
 import { useContext } from 'react'
 
-import ReservedTokens from '../Dashboard/ReservedTokens'
+import ReservedTokens from '../DashboardV1/ReservedTokens'
 import PayoutModsList from '../shared/PayoutModsList'
 import FundingCycleDetails from './FundingCycleDetails'
